# frontend/openapi.json
{"openapi":"3.0.1","info":{"title":"Swagger Hands-on API","description":"Spring Boot + Next.js Swagger ハンズオン用 REST API","contact":{"name":"Hands-on Team","email":"handson@example.com"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Local Development"}],"tags":[{"name":"Tasks","description":"タスク管理API"},{"name":"Users","description":"ユーザー管理API"}],"paths":{"/api/users/{id}":{"get":{"tags":["Users"],"summary":"ユーザー詳細取得","description":"指定したIDのユーザー情報を取得します","operationId":"getUserById","parameters":[{"name":"id","in":"path","description":"ユーザーID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"404":{"description":"ユーザーが見つかりません","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"取得成功","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["Users"],"summary":"ユーザー更新","description":"指定したIDのユーザー情報を更新します","operationId":"updateUser","parameters":[{"name":"id","in":"path","description":"ユーザーID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"404":{"description":"ユーザーが見つかりません","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"更新成功","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Users"],"summary":"ユーザー削除","description":"指定したIDのユーザーを削除します","operationId":"deleteUser","parameters":[{"name":"id","in":"path","description":"ユーザーID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"削除成功","content":{"*/*":{"schema":{"type":"object"}}}},"404":{"description":"ユーザーが見つかりません","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}}}},"/api/tasks/{id}":{"get":{"tags":["Tasks"],"summary":"タスク詳細取得","description":"指定したIDのタスク情報を取得します","operationId":"getTaskById","parameters":[{"name":"id","in":"path","description":"タスクID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"404":{"description":"タスクが見つかりません","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"取得成功","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["Tasks"],"summary":"タスク更新","description":"指定したIDのタスク情報を更新します","operationId":"updateTask","parameters":[{"name":"id","in":"path","description":"タスクID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}},"required":true},"responses":{"404":{"description":"タスクが見つかりません","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"更新成功","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Tasks"],"summary":"タスク削除","description":"指定したIDのタスクを削除します","operationId":"deleteTask","parameters":[{"name":"id","in":"path","description":"タスクID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"タスクが見つかりません","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"204":{"description":"削除成功","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users":{"get":{"tags":["Users"],"summary":"ユーザー一覧取得","description":"登録されている全ユーザーの一覧を取得します","operationId":"getAllUsers","responses":{"200":{"description":"取得成功","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"post":{"tags":["Users"],"summary":"ユーザー作成","description":"新しいユーザーを作成します","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"400":{"description":"バリデーションエラー","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"201":{"description":"作成成功","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/tasks":{"get":{"tags":["Tasks"],"summary":"タスク一覧取得","description":"タスクの一覧を取得します。ステータスや優先度でフィルタリング可能です。","operationId":"getAllTasks","parameters":[{"name":"status","in":"query","description":"ステータスでフィルタ","required":false,"schema":{"type":"string","description":"タスクのステータス","enum":["TODO","IN_PROGRESS","IN_REVIEW","DONE"]}},{"name":"priority","in":"query","description":"優先度でフィルタ","required":false,"schema":{"type":"string","description":"タスクの優先度","enum":["LOW","MEDIUM","HIGH","CRITICAL"]}}],"responses":{"200":{"description":"取得成功","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Task"}}}}}}},"post":{"tags":["Tasks"],"summary":"タスク作成","description":"新しいタスクを作成します","operationId":"createTask","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}},"required":true},"responses":{"400":{"description":"バリデーションエラー","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"201":{"description":"作成成功","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Task"}}}}}}},"/api/tasks/{id}/status":{"patch":{"tags":["Tasks"],"summary":"タスクステータス更新","description":"指定したIDのタスクのステータスのみを更新します","operationId":"updateTaskStatus","parameters":[{"name":"id","in":"path","description":"タスクID","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"status","in":"query","description":"新しいステータス","required":true,"schema":{"type":"string","description":"タスクのステータス","enum":["TODO","IN_PROGRESS","IN_REVIEW","DONE"]}}],"responses":{"404":{"description":"タスクが見つかりません","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"更新成功","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"User":{"required":["email","name"],"type":"object","properties":{"id":{"type":"integer","description":"ユーザーID","format":"int64","readOnly":true,"example":1},"name":{"maxLength":100,"minLength":1,"type":"string","description":"ユーザー名","example":"田中太郎"},"email":{"type":"string","description":"メールアドレス","example":"tanaka@example.com"}},"description":"タスクを管理するユーザー情報"},"ApiError":{"type":"object","properties":{"status":{"type":"integer","description":"HTTPステータスコード","format":"int32","example":400},"message":{"type":"string","description":"エラーメッセージ","example":"リクエストが不正です"},"detail":{"type":"string","description":"詳細情報","example":"タイトルは必須です"}},"description":"API共通エラーレスポンス"},"Task":{"required":["priority","status","title"],"type":"object","properties":{"id":{"type":"integer","description":"タスクID","format":"int64","readOnly":true,"example":1},"title":{"maxLength":200,"minLength":1,"type":"string","description":"タスクのタイトル","example":"Swagger定義を作成する"},"description":{"maxLength":1000,"minLength":0,"type":"string","description":"タスクの詳細説明","example":"OpenAPI 3.0仕様に準拠したSwagger定義を作成する"},"status":{"type":"string","description":"タスクのステータス","example":"TODO","enum":["TODO","IN_PROGRESS","IN_REVIEW","DONE"]},"priority":{"type":"string","description":"タスクの優先度","example":"HIGH","enum":["LOW","MEDIUM","HIGH","CRITICAL"]},"assigneeId":{"type":"integer","description":"担当ユーザーID","format":"int64","example":1},"createdAt":{"type":"string","description":"作成日時","format":"date-time","readOnly":true},"updatedAt":{"type":"string","description":"更新日時","format":"date-time","readOnly":true}},"description":"タスク情報"}}}}